# Optimise harmonic centrality