--- a/networkx/algorithms/centrality/harmonic.py
+++ b/networkx/algorithms/centrality/harmonic.py
@@ -80,10 +80,8 @@ def harmonic_centrality(G, nbunch=None, distance=None, sources=None):
     spl = partial(nx.shortest_path_length, G, weight=distance)
     for v in sources:
         dist = spl(v)
-        for u in nbunch.intersection(dist):
-            d = dist[u]
-            if d == 0:  # handle u == v and edges with 0 weight
-                continue
-            centrality[v if transposed else u] += 1 / d
+        for u, d_uv in dist.items():
+            if u in nbunch and d_uv != 0:
+                centrality[v if transposed else u] += 1 / d_uv
 
     return centrality