--- a/networkx/algorithms/centrality/harmonic.py
+++ b/networkx/algorithms/centrality/harmonic.py
@@ -81,7 +81,7 @@ def harmonic_centrality(G, nbunch=None, distance=None, sources=None):
     for v in sources:
         dist = spl(v)
         for u, d_uv in dist.items():
-            if u in nbunch and d_uv != 0:
+            if d_uv != 0 and u in nbunch:
                 centrality[v if transposed else u] += 1 / d_uv
 
     return centrality