@amcandio 2025-07-20
I see some variance on the measurements of asv. You should probably retry your profiling making sure your computer is not overloaded with other tasks